+++ pkg/backend/fanart/bilibili/notif.go
@@ -1,7 +1,6 @@
 package bilibili
 
 import (
-	"regexp"
 	"strconv"
 	"strings"
 
@@ -60,7 +59,7 @@ func (NotifData Notif) PushNotif(Color int) {
 		}
 
 		if tags == "_" && Group.NameGroup == "Independen" {
-
+			//do nothing,like my life
 		} else {
 			tmp, err := Bot.ChannelMessageSendEmbed(DiscordChannelID[i], engine.NewEmbed().
 				SetAuthor(strings.Title(Group.NameGroup), Group.IconURL).
@@ -76,13 +75,9 @@ func (NotifData Notif) PushNotif(Color int) {
 				SetColor(Color).MessageEmbed)
 			if err != nil {
 				log.Error(tmp, err.Error())
-				match, _ := regexp.MatchString("Unknown Channel", err.Error())
-				if match {
-					log.Info("Delete Discord Channel ", DiscordChannelID[i])
-					err = ChannelState.DelChannel()
-					if err != nil {
-						log.Error(err)
-					}
+				err = ChannelState.DelChannel(err.Error())
+				if err != nil {
+					log.Error(err)
 				}
 			}
 			err = engine.Reacting(map[string]string{