+++ pkg/backend/livestream/youtube/yt.go
@@ -2,13 +2,11 @@ package youtube
 
 import (
 	"math"
-	"math/rand"
 	"regexp"
 	"strings"
 	"sync"
 	"time"
 
-	config "github.com/JustHumanz/Go-simp/tools/config"
 	database "github.com/JustHumanz/Go-simp/tools/database"
 	engine "github.com/JustHumanz/Go-simp/tools/engine"
 	network "github.com/JustHumanz/Go-simp/tools/network"
@@ -34,7 +32,6 @@ func CheckSchedule() {
 					"Vtube Region": Member.Region,
 				}).Info("Checking Youtube")
 				go StartCheckYT(Member, Group, &wg)
-				time.Sleep(time.Duration(rand.Intn(config.RandomSleep-400)+400) * time.Millisecond)
 			}
 		}
 		wg.Wait()