+++ pkg/frontend/discord_handler/general_handler.go
@@ -3,6 +3,7 @@ package discordhandler
 import (
 	"encoding/json"
 	"fmt"
+	"math/rand"
 	"regexp"
 	"strconv"
 	"strings"
@@ -1198,6 +1199,11 @@ func EnableState(s *discordgo.Session, m *discordgo.MessageCreate) {
 	m.Content = strings.ToLower(m.Content)
 	Prefix := config.PGeneral
 	if strings.HasPrefix(m.Content, Prefix) {
+		Color, err := engine.GetColor("/tmp/discordpp.tmp", m.Author.AvatarURL("128"))
+		if err != nil {
+			log.Error(err)
+		}
+
 		var (
 			ChannelState = database.DiscordChannel{
 				TypeTag:     0,
@@ -1207,8 +1213,31 @@ func EnableState(s *discordgo.Session, m *discordgo.MessageCreate) {
 			}
 			already []string
 			done    []string
-			msg1    = "Remember boys,always respect the author,**do not save the fanart without permission from the author**"
-			msg2    = "Every livestream have some **rule**,follow the **rule** and don't be asshole"
+			msg1    = engine.NewEmbed().
+				SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
+				SetDescription("Remember boys,always respect the author,**do not save the fanart without permission from the author**").
+				SetThumbnail(config.GoSimpIMG).
+				SetColor(Color)
+
+			msg2 = engine.NewEmbed().
+				SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
+				SetDescription("Every livestream have some **rule**,follow the **rule** and don't be asshole").
+				SetThumbnail(config.GoSimpIMG).
+				SetColor(Color)
+
+			msg3 = engine.NewEmbed().
+				SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
+				SetDescription("Remember boys,always respect the author,**do not save the fanart without permission from the author**\nEvery livestream have some **rule**,follow the **rule** and don't be asshole").
+				SetThumbnail(config.GoSimpIMG).
+				SetColor(Color)
+
+			msg4 = engine.NewEmbed().
+				SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
+				SetTitle("Donate").
+				SetDescription("Buy dev coffie to improve bot performance and make dev happy").
+				SetURL(config.KoFiLink).
+				SetThumbnail(config.GoSimpIMG).
+				SetColor(Color)
 		)
 		CommandArray := strings.Split(m.Content, " ")
 		if CommandArray[0] == Prefix+Enable {
@@ -1301,18 +1330,42 @@ func EnableState(s *discordgo.Session, m *discordgo.MessageCreate) {
 					if err != nil {
 						log.Error(err)
 					}
+					msgID := ""
 					if ChannelState.TypeTag == 1 {
-						_, err := s.ChannelMessageSend(m.ChannelID, msg1+"\n")
+						tmp, err := s.ChannelMessageSendEmbed(m.ChannelID, msg1.MessageEmbed)
 						if err != nil {
 							log.Error(err)
 						}
+						msgID = tmp.ID
 					} else if ChannelState.TypeTag == 2 {
-						_, err := s.ChannelMessageSend(m.ChannelID, msg2+"\n")
+						tmp, err := s.ChannelMessageSendEmbed(m.ChannelID, msg2.MessageEmbed)
 						if err != nil {
 							log.Error(err)
 						}
+						msgID = tmp.ID
 					} else {
-						_, err := s.ChannelMessageSend(m.ChannelID, msg1+"\n"+msg2+"\n")
+						tmp, err := s.ChannelMessageSendEmbed(m.ChannelID, msg3.MessageEmbed)
+						if err != nil {
+							log.Error(err)
+						}
+						msgID = tmp.ID
+					}
+					MessagePinned, err := s.ChannelMessagesPinned(m.ChannelID)
+
+					for _, Message := range MessagePinned {
+						if Message.Author.ID == BotInfo.ID {
+							err := s.ChannelMessageUnpin(m.ChannelID, Message.ID)
+							if err != nil {
+								log.Error(err)
+							}
+						}
+					}
+					err = s.ChannelMessagePin(m.ChannelID, msgID)
+					if err != nil {
+						log.Error(err)
+					}
+					if rand.Float32() < 0.5 && config.KoFiLink != "" {
+						_, err := s.ChannelMessageSendEmbed(m.ChannelID, msg4.MessageEmbed)
 						if err != nil {
 							log.Error(err)
 						}
@@ -1344,7 +1397,7 @@ func EnableState(s *discordgo.Session, m *discordgo.MessageCreate) {
 					if CheckPermission(m.Author.ID, m.ChannelID, s) {
 						ChannelState.SetVtuberGroupID(VTuberGroup.ID)
 						if ChannelState.ChannelCheck() {
-							err := ChannelState.DelChannel()
+							err := ChannelState.DelChannel("Delete")
 							if err != nil {
 								log.Error(err)
 								_, err := s.ChannelMessageSend(m.ChannelID, "Something error XD")
@@ -1354,7 +1407,6 @@ func EnableState(s *discordgo.Session, m *discordgo.MessageCreate) {
 								return
 							}
 							done = append(done, "`"+VTuberGroup.NameGroup+"`")
-
 						} else {
 							already = append(already, "`"+VTuberGroup.NameGroup+"`")
 						}
@@ -1514,22 +1566,48 @@ func EnableState(s *discordgo.Session, m *discordgo.MessageCreate) {
 					if err != nil {
 						log.Error(err)
 					}
+
+					msgID := ""
 					if ChannelState.TypeTag == 1 {
-						_, err := s.ChannelMessageSend(m.ChannelID, msg1+"\n")
+						tmp, err := s.ChannelMessageSendEmbed(m.ChannelID, msg1.MessageEmbed)
 						if err != nil {
 							log.Error(err)
 						}
+						msgID = tmp.ID
 					} else if ChannelState.TypeTag == 2 {
-						_, err := s.ChannelMessageSend(m.ChannelID, msg2+"\n")
+						tmp, err := s.ChannelMessageSendEmbed(m.ChannelID, msg2.MessageEmbed)
 						if err != nil {
 							log.Error(err)
 						}
+						msgID = tmp.ID
 					} else {
-						_, err := s.ChannelMessageSend(m.ChannelID, msg1+"\n"+msg2+"\n")
+						tmp, err := s.ChannelMessageSendEmbed(m.ChannelID, msg3.MessageEmbed)
+						if err != nil {
+							log.Error(err)
+						}
+						msgID = tmp.ID
+					}
+					MessagePinned, err := s.ChannelMessagesPinned(m.ChannelID)
+
+					for _, Message := range MessagePinned {
+						if Message.Author.ID == BotInfo.ID {
+							err := s.ChannelMessageUnpin(m.ChannelID, Message.ID)
+							if err != nil {
+								log.Error(err)
+							}
+						}
+					}
+					err = s.ChannelMessagePin(m.ChannelID, msgID)
+					if err != nil {
+						log.Error(err)
+					}
+					if rand.Float32() < 0.5 && config.KoFiLink != "" {
+						_, err := s.ChannelMessageSendEmbed(m.ChannelID, msg4.MessageEmbed)
 						if err != nil {
 							log.Error(err)
 						}
 					}
+
 				} else {
 					_, err := s.ChannelMessageSend(m.ChannelID, strings.Join(already, ",")+" Same state")
 					if err != nil {