+++ pkg/guild/guild_handler.go
@@ -55,15 +55,28 @@ func GuildJoin(s *discordgo.Session, g *discordgo.GuildCreate) {
 					return
 				}
 				if Channel.Type == 0 && BotPermission&2048 != 0 {
-					s.ChannelMessageSendEmbed(Channel.ID, engine.NewEmbed().
-						SetTitle("Thx for invite me to this server <3 ").
-						SetURL("https://go-simp.human-z.tech/Guide/").
-						SetThumbnail(config.GoSimpIMG).
-						SetImage(H3llcome[rand.Intn(len(H3llcome))]).
-						SetColor(14807034).
-						AddField("Setup", "You can watch [here](https://go-simp.human-z.tech/Guide/)").
-						AddField("Need support?", "Join [dev server](https://discord.com/invite/ydWC5knbJT)").
-						InlineAllFields().MessageEmbed)
+					if config.KoFiLink == "" {
+						s.ChannelMessageSendEmbed(Channel.ID, engine.NewEmbed().
+							SetTitle("Thx for invite me to this server <3 ").
+							SetURL("https://go-simp.human-z.tech/Guide/").
+							SetThumbnail(config.GoSimpIMG).
+							SetImage(H3llcome[rand.Intn(len(H3llcome))]).
+							SetColor(14807034).
+							AddField("Setup", "You can watch [here](https://go-simp.human-z.tech/Guide/)").
+							AddField("Need support?", "Join [dev server](https://discord.com/invite/ydWC5knbJT)").
+							InlineAllFields().MessageEmbed)
+					} else {
+						s.ChannelMessageSendEmbed(Channel.ID, engine.NewEmbed().
+							SetTitle("Thx for invite me to this server <3 ").
+							SetURL("https://go-simp.human-z.tech/Guide/").
+							SetThumbnail(config.GoSimpIMG).
+							SetImage(H3llcome[rand.Intn(len(H3llcome))]).
+							SetColor(14807034).
+							AddField("Setup", "You can watch [here](https://go-simp.human-z.tech/Guide/)").
+							AddField("Need support?", "Join [dev server](https://discord.com/invite/ydWC5knbJT)").
+							InlineAllFields().
+							AddField("Support "+BotID.Username, "[Ko-Fi]("+config.KoFiLink+")").MessageEmbed)
+					}
 
 					//Save discord name to database
 					err := DataGuild.InputGuild()